+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Add optional `voteLabel` prop to the `VoteProposalDataListItem` component in order to pass a custom string wen
+-   Add optional `confirmationLabel` prop to the `VoteProposalDataListItem` component in order to pass a custom string when
     needed.
 -   Add new `xs` size for core `Avatar` component